Change 'score' field from Long to Integer

+++ src/org/dyndns/doujindb/plug/impl/dataimport/Metadata.java
@@ -19,7 +19,7 @@ abstract class Metadata
 	Metadata() { }
 
 	@XmlElement(name="score")
-	protected Long score;
+	protected Integer score;
 	
 	@XmlElement(name="name")
 	protected String name;
