+++ src/org/dyndns/doujindb/plug/impl/dataimport/MugiMugiProvider.java
@@ -63,7 +63,7 @@ final class MugiMugiProvider extends MetadataProvider {
 			// Generate Metadata
 			Metadata md = toMetadata(book);
 			// Set 'score' field
-			md.score = (long) bestMatch;
+			md.score = (int) bestMatch;
 			// Generate Warning in case of poor similarity results
 			if(bestMatch < Configuration.provider_mugimugi_threshold.get()) {
 				String message = "Response book did not match the threshold (" + bestMatch + "%)";
